# Conflicts:
#	src/server/TrackingServerImpl.java

+++ src/peer/FileDownloadBundle.java
@@ -5,5 +5,5 @@ import java.io.Serializable;
 public class FileDownloadBundle implements Serializable {
     public String fileName;
     public byte[] fileContents;
-    public int checksum;
+    public long checksum;
 }
